% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoranger.R
\name{extract_fusion_gene}
\alias{extract_fusion_gene}
\title{extract fusion genes or CAR sequences}
\usage{
extract_fusion_gene(
  BC.data.file,
  WILDTYPE,
  FUSION,
  downsample = NA,
  filter = 4
)
}
\arguments{
\item{BC.data.file}{output of fusion_gene.py}

\item{WILDTYPE}{wildtype gene}

\item{FUSION}{fusion partner}

\item{downsample}{maximum number of reads to process}

\item{FILTER}{minimum number of reads per cell barcode}
}
\value{
dataframe with cell barcodes and information on full length (wildtype) or truncated transcript (mutated)
}
\description{
extract fusion genes or CAR sequences
}
