% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoranger.R
\name{extract_indel}
\alias{extract_indel}
\title{extract indels or insertions from}
\usage{
extract_indel(BC.data.file, REF, downsample = NA, CONSENSUS = -1, FILTER = 4)
}
\arguments{
\item{BC.data.file}{output of call_mutations.py}

\item{REF}{reference allele (A, C, G, T)}

\item{downsample}{maximum number of reads to process}

\item{CONSENSUS}{number of nucleotides deleted or inserted}

\item{FILTER}{minimum number of reads per cell barcode}
}
\value{
dataframe with cell barcodes and insertion/deletion
}
\description{
extract indels or insertions from
}
