% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoranger.R
\name{extract_isoforms}
\alias{extract_isoforms}
\title{extract reads covering exon}
\usage{
extract_isoforms(BC.data.file, GENE, EXON, downsample = NA, filter = 4)
}
\arguments{
\item{BC.data.file}{output of fusion_gene.py}

\item{GENE}{gene}

\item{EXON}{exon}

\item{downsample}{maximum number of reads to process}

\item{FILTER}{minimum number of reads per cell barcode}
}
\value{
dataframe with cell barcodes and information on exon expression
}
\description{
extract reads covering exon
}
