% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoranger.R
\name{extract_length_diff}
\alias{extract_length_diff}
\title{detect truncated transcripts}
\usage{
extract_length_diff(BC.data.file, POS, downsample = NA, FILTER = 4)
}
\arguments{
\item{BC.data.file}{output of fusion_gene.py}

\item{POS}{position where truncation occurs}

\item{downsample}{maximum number of reads to process}

\item{FILTER}{minimum number of reads per cell barcode}

\item{CONSENSUS}{number of nucleotides deleted or inserted}
}
\value{
dataframe with cell barcodes and information on full length (wildtype) or truncated transcript (mutated)
}
\description{
detect truncated transcripts
}
