% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoranger.R
\name{extract_mutation}
\alias{extract_mutation}
\title{extract single nucleotide variants from piled up reads generated by call_mutations.py}
\usage{
extract_mutation(BC.data.file, ALT, REF, downsample = NA, FILTER = 4)
}
\arguments{
\item{BC.data.file}{output of call_mutations.py}

\item{ALT}{alternative allele (A, C, G, T)}

\item{REF}{reference allele (A, C, G, T)}

\item{downsample}{maximum number of reads to process}

\item{FILTER}{minimum number of reads per cell barcode}
}
\value{
dataframe with cell barcodes and mutation
}
\description{
extract single nucleotide variants from piled up reads generated by call_mutations.py
}
