% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoranger.R
\name{knee_plot}
\alias{knee_plot}
\title{plot knee plot}
\usage{
knee_plot(x, cutoff = NA, downsample = NA)
}
\arguments{
\item{x}{pileup data}

\item{cutoff}{draw cut off at indicated number of reads}

\item{downsample}{downsample to indicated number of reads}
}
\description{
plot knee plot
}
